file restructure and additional architecture work
I have made additions to the renderer, it is about to draw quads now.

A key decision I have made is to split each screen into a class, which interfaces with IGameScreen. This is to encapsule rendering and logic code within their separate areas without creating a chain of if statemnets inside of MainWindow when deciding what state the game is in. The instance of IGameScreen will reflect what state the game is in and go down that path of logic.

I will need to create a signal of some kind, maybe an event, that triggers when the current screen needs to change, what screen it needs to change to, and what information needs to be passed into that screen (chart data, replay data as examples).

The shader has been expanded to include some more functionality:
- draw only a quad thats coloured
- draw a textured quad
- draw a textured quad with a colour tint on top (needed for later, such as for reflecting UI interactions and for showing if a long note was dropped before the tail was judged.

Orchestrator has been renamed to GameSession.

+++ engine/Note.cs
@@ -5,8 +5,8 @@
         public NoteType noteType { get; set; }
         public NoteState noteState { get; set; }
         public byte column { get; set; }
-        public double startTime { get; set; }
-        public double endTime { get; set; }
+        public float startTime { get; set; }
+        public float endTime { get; set; }
         public float headPosY { get; set; }
         public float tailPosY { get; set; }
     }
